Test log and invalid properties
